Add Lightstep bootstrapping and configuration
For our December POC.

Lightstep uses gRPC behind the scenes, which in its out-of-the-box form requires ALPN support.  KairosDB itself uses Jetty and requires the jetty-alpn extension to be on the boot classpath.  To get the right version we can use the jetty-alpn-agent.  The pom was updated to ensure its presence, and the KairosDB startup script updated to use it.  This will require us to maintain ALPN versioning infomation in two locations, pom.xml, and src/scripts/kairosdb.sh.

See:
- https://github.com/grpc/grpc-java/blob/master/SECURITY.md#tls-with-jdk-jetty-alpnnpn
- https://github.com/grpc/grpc-java/blob/master/SECURITY.md#netty

The rest of the changes are straightforward.  A configuration class for essential Lightstep values was added, and the default properties updated with empty values.  This is required or Guice will not attempt to populate the configuration class with provided environment variables.  These values include the Lightstep API key so should not ever be provided in the properties file itself.

Finally, the Main class was updated to register a global tracer if both an API key and a collector host were provided in the environment.  Port and protocol have defaults of https and 443, but can be overriden.  A tracer will not be registered if either the API key or collector host is not present.

Tested against a local dockerized Cassandra.