Provide metric specific index order config
For some cases the default order is not the optimal one to fetch data
from Cassandra. This PR adds a config to be able to change the default
order of indexes.
The default order is set in the `kairosdb.datastore.cassandra.index_tag_list`
config setting. The new setting can be set in
`kairosdb.datastore.cassandra.metric_index_tag_list`. Format of this one
is `metric1=key1,key2,key3;metric2=key2,key1,key3`.